+++ R/arguments.R
@@ -67,7 +67,8 @@ set_args <- function(object, ...) {
     eng_args = object$eng_args,
     mode = object$mode,
     method = NULL,
-    engine = object$engine
+    engine = object$engine,
+    check_missing_spec = FALSE
   )
 }