+++ R/mlp.R
@@ -64,7 +64,7 @@ print.mlp <- function(x, ...) {
   cat("Single Layer Neural Network Specification (", x$mode, ")\n\n", sep = "")
   model_printer(x, ...)
 
-  if(!is.null(x$method$fit$args)) {
+  if (is_printable_spec(x)) {
     cat("Model fit template:\n")
     print(show_call(x))
   }