+++ R/poisson_reg.R
@@ -56,7 +56,7 @@ print.poisson_reg <- function(x, ...) {
   cat("Poisson Regression Model Specification (", x$mode, ")\n\n", sep = "")
   model_printer(x, ...)
 
-  if (!is.null(x$method$fit$args)) {
+  if (is_printable_spec(x)) {
     cat("Model fit template:\n")
     print(show_call(x))
   }