+++ R/proportional_hazards.R
@@ -58,7 +58,7 @@ print.proportional_hazards <- function(x, ...) {
   cat("Proportional Hazards Model Specification (", x$mode, ")\n\n", sep = "")
   model_printer(x, ...)
 
-  if (!is.null(x$method$fit$args)) {
+  if (is_printable_spec(x)) {
     cat("Model fit template:\n")
     print(show_call(x))
   }