+++ tests/testthat/test_args_and_modes.R
@@ -77,7 +77,9 @@ test_that("unavailable modes for an engine and vice-versa", {
   )
 
   expect_error(
-    proportional_hazards() %>% set_mode("regression"),
+    expect_message(
+      proportional_hazards() %>% set_mode("regression")
+    ),
     "'regression' is not a known mode"
   )
 
@@ -92,7 +94,9 @@ test_that("unavailable modes for an engine and vice-versa", {
   )
 
   expect_error(
-    proportional_hazards() %>% set_engine(),
+    expect_message(
+      proportional_hazards() %>% set_engine()
+    ),
     "No known engines for"
   )
 })