+++ tests/testthat/test_proportional_hazards.R
@@ -1,7 +1,6 @@
 test_that("printing", {
-  expect_output(
-    print(proportional_hazards()),
-    "Proportional Hazards Model Specification \\(censored regression\\)"
+  expect_snapshot(
+    proportional_hazards()
   )
 })
 
@@ -19,7 +18,9 @@ test_that("bad input", {
 
 test_that("wrong fit interface", {
   expect_error(
-    proportional_hazards() %>% fit_xy(),
+    expect_message(
+      proportional_hazards() %>% fit_xy()
+    ),
     "must use the formula interface"
   )
 })